requirements.txt
autopep8==1.4.4
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
paho-mqtt==1.5.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.1
pythonnet==2.4.0
pywebview==3.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
SQLAlchemy==1.3.11
uritemplate==3.0.0
urllib3==1.25.7
websocket-server==0.4
Werkzeug==0.16.0
